Fix incorrect SunCalc day
When the time ticked past 12am and into a new day, it triggered
a recalculation of the sun times. However, the timestamp being
passed to SunCalc was interpreted as the previous day, which
led to negative numbers being displayed after sunrise.

This ensures the Date object passed to SunCalc is squarely in
the middle of the day, resulting in a correct date being chosen.